+++ lab5/Editor/Document/DocumentItems/Elements/Image/Image.cpp
@@ -23,7 +23,7 @@ size_t Image::GetHeight() const
 	return m_height;
 }
 
-void Image::Resize(int width, int height)
+void Image::Resize(size_t width, size_t height)
 {
 	m_width = width;
 	m_height = height;
