+++ lab5/Editor/Document/IDocument.h
@@ -24,7 +24,7 @@ public:
 	// Параметр path задает путь к вставляемому изображению
 	// При вставке изображение должно копироваться в подкаталог images
 	// под автоматически сгенерированным именем
-	virtual std::shared_ptr<IImage> InsertImage(const Path& path, int width, int height,
+	virtual std::shared_ptr<IImage> InsertImage(const Path& path, size_t width, size_t height,
 		std::optional<size_t> position = std::nullopt)
 		= 0;
